// src/app/protected/interfaces/GaranteReferencia.ts
export interface GaranteReferencia {

  codGarante?:         number;
  codPersona?:         number;
  codEmpleado?:        number;
  direccionTrabajo?:   string;
  empresaTrabajo?:     string;
  tipo?:               string;
  observacion?:        string;
  audUsuario?:         number;

  esEmpleado?:         string;
  nombreCompleto?:     string;
  ok?:                 string;
  msg?:                string;
  direccionDomicilio?: string;
}